legacy/requirements.txt
gradio==3.49.0
selenium
beautifulsoup4
requests
chromedriver-autoinstaller
safetensors
pyperclip
transformers
pydantic
pillow
torch
sentence-transformers
gensim
fasttext-wheel
numpy
gdown
discord.py
jishaku
psutil
GPUtil
py-cpuinfo
pynvml
pypresence
nvidia-ml-py
tqdm